# Spanify Cryptography.Base64Transform