## Description

`Convert.{To|From}Base64CharArray` was used, so this involved temporary buffer allocations and copying between them. With https://github.com/dotnet/corefx/blob/adc16f3d53b18467d5ed92fd1b988b0c21901ab9/src/System.Memory/ref/System.Memory.cs#L397 we have a nice API that does the same, but based on spans. So the code got rewritten to use the span-based conversion, thus saving allocation and unnecessary copying.

## Benchmarks

[Source](https://github.com/gfoidl/Benchmarks/tree/f0ed5a11f9a14b04f90f0a643d34ef8f39bc63c6/corefx/System/Security/Cryptography/Base64TransformsBenchmarks/Base64TransformsBenchmarks/Benchmarks)

### ToBase64Transform

#### TransformBlock

```
|   Method |     Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |---------:|----------:|----------:|------:|-------:|------:|------:|----------:|
| Original | 85.65 ns | 0.4726 ns | 0.4189 ns |  1.00 | 0.0101 |     - |     - |      64 B |
|       PR | 30.24 ns | 0.3475 ns | 0.3080 ns |  0.35 |      - |     - |     - |         - |
```

#### TransformFinalBlock

```
|   Method |     Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |---------:|----------:|----------:|------:|-------:|------:|------:|----------:|
| Original | 74.00 ns | 0.7205 ns | 0.6387 ns |  1.00 | 0.0101 |     - |     - |      64 B |
|       PR | 43.51 ns | 0.4889 ns | 0.4334 ns |  0.59 | 0.0051 |     - |     - |      32 B |
```

### FromBase64Transform

#### TransformBlock

```
|   Method |      Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |----------:|----------:|----------:|------:|-------:|------:|------:|----------:|
| Original | 199.08 ns | 1.0918 ns | 1.0212 ns |  1.00 | 0.0203 |     - |     - |     128 B |
|       PR |  83.15 ns | 0.7017 ns | 0.6564 ns |  0.42 |      - |     - |     - |         - |

```

#### TransformFinalBlock

```
|   Method |      Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |----------:|----------:|----------:|------:|-------:|------:|------:|----------:|
| Original | 207.60 ns | 2.1021 ns | 1.8635 ns |  1.00 | 0.0203 |     - |     - |     128 B |
|       PR |  97.46 ns | 0.6792 ns | 0.6353 ns |  0.47 | 0.0050 |     - |     - |      32 B |

```