@gfoidl 2019-06-06
@bartonjs now I understand, that makes sense in this context. I'll address these points over the weekend (quite busy right now).