@bartonjs 2019-06-19
> D:\a\1\s\.packages\microsoft.net.compilers.toolset\3.2.0-beta3-19304-07\tasks\netcoreapp2.1\Microsoft.CSharp.Core.targets(59,5): error MSB3105: The item "D:\a\1\s\src\Common\src\System\Security\Cryptography\CryptoPool.cs" was specified more than once in the "Sources" parameter.  Duplicate items are not supported by the "Sources" parameter. [D:\a\1\s\src\System.Security.Cryptography.Encoding\src\System.Security.Cryptography.Encoding.csproj]

Looks like there already was one on line 104 (for OSX only).  You can just delete that one now, since it's now unconditionally included.