@gfoidl 2019-06-19
> for inclusion in .NET Core 3.0 

@bartonjs FYI I'll update this PR on Friday (maybe tomorrow, but can't promise).