@gfoidl 2019-06-21
## Benchmarks

[Source](https://github.com/gfoidl/Benchmarks/tree/e86a5d34769738377b0090a03b94ba9bd0b691e3/corefx/System/Security/Cryptography/Base64TransformsBenchmarks/Base64TransformsBenchmarks/Benchmarks)

Based on latest commit (4215210f249a17b2ed776ea66616e187958ba176).

### ToBase64Transform

#### TransformBlock

```
|   Method |     Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |---------:|----------:|----------:|------:|-------:|------:|------:|----------:|
| Original | 84.18 ns | 0.7220 ns | 0.6754 ns |  1.00 | 0.0021 |     - |     - |      64 B |
|       PR | 29.67 ns | 0.1367 ns | 0.1212 ns |  0.35 |      - |     - |     - |         - |
```

#### TransformFinalBlock

```
|   Method |     Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |---------:|----------:|----------:|------:|-------:|------:|------:|----------:|
| Original | 71.26 ns | 0.2582 ns | 0.2156 ns |  1.00 | 0.0021 |     - |     - |      64 B |
|       PR | 41.39 ns | 0.2874 ns | 0.2548 ns |  0.58 | 0.0011 |     - |     - |      32 B |
```

### FromBase64Transform

#### TransformBlock

```
|   Method |     Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |---------:|----------:|----------:|------:|-------:|------:|------:|----------:|
| Original | 197.2 ns | 0.7589 ns | 0.6728 ns |  1.00 | 0.0043 |     - |     - |     128 B |
|       PR | 156.2 ns | 0.7158 ns | 0.6695 ns |  0.79 |      - |     - |     - |         - |

```

#### TransformFinalBlock

```
|   Method |      Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |----------:|----------:|----------:|------:|-------:|------:|------:|----------:|
| Original | 210.18 ns | 1.9161 ns | 1.6985 ns |  1.00 | 0.0043 |     - |     - |     128 B |
|       PR |  94.82 ns | 0.3535 ns | 0.3134 ns |  0.45 | 0.0011 |     - |     - |      32 B |

```