@bartonjs 2019-06-24
It looks reasonable to me; but I'd like a second set of eyes (@stephentoub).