@gfoidl 2019-07-09
## Benchmarks

[Source](https://github.com/gfoidl/Benchmarks/tree/5bd9f20becf37dc11532ebe01ab003d90f118658/corefx/System/Security/Cryptography/Base64TransformsBenchmarks/Base64TransformsBenchmarks/Benchmarks)

Based on latest commit (3dcedd521493c0552d6b82021815647530d00e3d).

### ToBase64Transform

Didn't change from [last results](https://github.com/dotnet/corefx/pull/38062#issuecomment-504578089).

### FromBase64Transform

#### TransformBlock

```
|   Method |      Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |----------:|----------:|----------:|------:|-------:|------:|------:|----------:|
| Original | 204.66 ns | 1.0553 ns | 0.9355 ns |  1.00 | 0.0043 |     - |     - |     128 B |
|       PR |  85.59 ns | 0.4166 ns | 0.3693 ns |  0.42 |      - |     - |     - |         - |

```

#### TransformFinalBlock

```
|   Method |      Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |----------:|----------:|----------:|------:|-------:|------:|------:|----------:|
| Original | 196.32 ns | 1.7872 ns | 1.6718 ns |  1.00 | 0.0043 |     - |     - |     128 B |
|       PR |  99.09 ns | 0.4190 ns | 0.3715 ns |  0.51 | 0.0011 |     - |     - |      32 B |

```